New material
Basis for a new material that support the metal-roughness workflow.
The basic idea behind this material is to mimic the principled shader in
blender, aka the Disney principled shader, where the main goal is
to have a relatively simplistic material that covers most use-cases.
Not all properties of the original principled materials are available,
as they are not considered relevant, yet.

Keep in mind that this is still a bit rough and is built on-top of the
existing default specular material. Further improvements are needed
to make the materials be more one-to-one with the material designer
tools.

Task-number: QTBUG-76690
Change-Id: I2ead392bbf1c1b68ce13e64b5f88ab52246a4947
Reviewed-by: Andy Nichols <andy.nichols@qt.io>