Add Loader3D Component
The Loader component from QtQuick is a subclass of QQuickItem, and can
not be easily used to load components inside of a 3D scene.  Loader3D is
a subclass of Node so it can be used in a spatial scene.

This class should be replaced by a unified Loader component in Qt 6.

Change-Id: Id859116c47dad4e67e757bd2d30ce30cf8469415
Reviewed-by: Christian Strømme <christian.stromme@qt.io>