QQuick3DNode: only calculate global transform when changed
Since we cannot wait for the backend QSSGRenderNode to update our
global transform whenever queried by the app, we sometimes
need to calculate it immediately in the front end as well.
However, currently we recalculate it every time someone asks for
it (e.g if the app calls node.globalPosition followed by
node.globalRotation, we end up calculating it twice).

This patch will add a m_dirtyGlobalTransform flag that gets updated
whenever a node, or an ancestor of the node, changes geometry. That
way we only calculate the global transform once (and only when
someone asks for it).

Note, currently we end up duplicating and calculating the global
transform _for some nodes_ in the frontend and the backend.
In the future, it makes sense to investigate how these two can
be merged.

Change-Id: I4630d1f1c7431a77e866f9c6625b825eb3d66941
Reviewed-by: Andy Nichols <andy.nichols@qt.io>