QQuick3DNode: use 'scene' instead of 'global'
The word 'global' is used all over Qt to mean "relative to the
screen". E.g QQuickItem has mapping functions "mapToGlobal" and
"mapFromGlobal", which basically maps a position to and from
screen space. The same with QMouseEvent, it has a globalPos()
function that gives you the mouse position in screen space.

In QtQuick3D, we understand "global space" to be the root
of the scene. In other words, we use "global" differently
than the rest of Qt. E.g 'Node.globalPosition' gives
you a nodes position in scene space, not screen space.

This is a source for confusion when a developer starts mixing
QtQuick, and QtQuick3D (along with QWindow and QEvents in general).

This patch will rename some of the properties and functions that
uses the word 'global' to instead use 'scene'. This makes the API
more coherent with QtQuick and the rest of Qt.

Change-Id: I8044230fbc5452f1739cb6a6abb2d3d2eb63f5af
Reviewed-by: Andy Nichols <andy.nichols@qt.io>