Refactor OptionsManager so that we can support building with 5.12
Qt Design Studio is currently built against 5.12, but is building using
the 5.14 module of Qt Quick3D.  In the mean time there was no default
constructor for QCommandLineOptions, so it's not possible to return a
QVector of QCommandLineOption values.

So I cleaned this up a bit and just have a table to heap based
QCommandLineOptions. This should work with 5.12 and greater and is a bit
cleaner.

Change-Id: Ic0a16d7f094ce685cfadf31ec8960f14496f69cc
Reviewed-by: Christian Strømme <christian.stromme@qt.io>