Always render when calling QSSGLayerRenderData::runnableRenderToViewport
It doesn't make sense to try and drop out early from rendering here,
because with the current setup path setup, if we get into this function
we know that something is dirty in the scene already.  Also the logic
didn't work when the background mode was Transparent and there were not
objects in the scene.  You still need to clear with transparent in that
case to remove any old content from the scene.

Fixes: QTBUG-79104
Change-Id: I44d0f8a5ca2588ce1ce0cdaedf16e591798296b0
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>