Make it functional when Quick is running on QRhi with the OpenGL backend
Running a Quick 3D application with QSG_RHI=1 QSG_RHI_BACKEND=gl is
expected to work, even though Qt Quick is not going directly to OpenGL
in this setup.

Start using the new 5.14 APIs as appropriate. All View3D rendering
modes should now work both on the default (direct OpenGL) and the
QRhi-based rendering paths of Qt Quick.

We take one shortcut: we still connect to before/afterRendering() when
Quick is not using the RHI, even though this should not be necessary
strictly speaking.

This is because Quick3D trashes the state in some unforeseen ways (it
is something with vertex array objects), which then breaks the Quick
renderer when the Quick3D rendering is hooked up to
beforeRenderPassRecording() (which is a signal that gets emitted from
within the Quick rendering process, unlike beforeRendering() which is
emitted a lot earlier). The rhi-based rendering path deals better with
this due to the QRhi backend centrally controlling the OpenGL state.

Task-number: QTBUG-79221
Change-Id: I03e3ef3561b83c2b7281d05da018df39349f11d3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>