Improve shader key hashing
This improves the time spent hashing the shader key to 1/20th of the
previously measured time. This is done by hashing the shader defines
up-front and storing them in the QSSGShaderPreprocessorFeature class.
In addition to storing the keys we also now store the name of the define
as a char pointer, further improving the hashing, as the pointer should
already be pointing to a unique location and therefore allow us to take
the fast path.

Change-Id: I4eb871345e09f456c19a1ff559b38c5b473ba1f3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>