Material: Remove the emissivePower property from the principled material
A uniform emissive factor can be applied on the front-end by the user
through the emissveColor property if needed. The emissveColor property
also doubles as the factor used for the RGB components when an emissive
map is used, improving flexibility.
This change also moves the object opacity to the material_property
instead of keeping it in the material_diffuse uniform.

Fixes: QTBUG-79467
Change-Id: I3a4cb282528eff8815fdf0ac68ced86edde6cce2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Antti Määttä <antti.maatta@qt.io>