Implement custom geometry
Add geometry property to model. The geometry is abstract base class,
with which the user can implement custom geometries.

The custom geometry needs to use the QSSGRenderGeometry class to add the
vertices, indices, attributes and other properties. When the geometry is
set and the model doesn't have source path it will use the geometry
as a mesh.

Task-number: QTBUG-79733
Change-Id: I0d1a728b9092790cd7e690a8d37e6df531584a86
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>