Remove OffscreenRenderer dead code
This code path was used to render sub-presentations. I used this code
path fairly on to provide "offscreen" rendering, but it is largely
unnecessary and is not that great of an interface to begin with.  We
still support offscreen rendering, but do so by adding extra logic in
the "direct" rendering path which is easier to maintain.  This just
removes all of the dead code.

Change-Id: Iff6fda054a3dbdeecdaf03ec273dd2a96f3ca378
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>