Remove ReferencedMaterial dead code
This was a leftover from Qt 3D Studio. There really is no reason to
reference materials on the backend since the engine does such a good job
of reusing materials with the same shader key anyway.  On the frontend
material referencing is already implimented, and we never used this
backend node.

Change-Id: I57190585e7c52604405ff09522fcb8b55f63263c
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>