Clean up Renderer code
There is still a lot of leftover dead code from Qt 3D Studio's renderer
that needed to be removed to simplify renderer code.  Much of this stems
from the composition engine that Qt 3D Studio provided that we avoid and
instead use Qt Quick for.  This means removing everything related to
scaling factor, presentation vs window vs layer sizing, and the concept
of rendering to a texture.

The paths for rendering to a texture are uncessary with the new
approach, because we still can render to a texture, we just just have
any "sub-passes" in the 3D renderer.  Instead we use the same approach
that the Qt Quick renderer uses with offscreen targets, and spawn a new
instance of the renderer for each subpass and instead target an
offscreen render target.  Qt Quick provides the "render graph" mechanism
that makes sure offscreen subpass dependeicnes are rendered in the
correct order.

By cleaning up all of these things, it should make it more clear what
code paths are alive, and make it easier to adjust the rendering
strategy needed to make the RHI port possible.

Change-Id: I6f1bbddd90f585f2812e26a9d9e7e26c583ab676
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>