Add spot light
Add a spot light type. It has the attenuation properties from point
light and additionally coneAngle and innerConeAngle properties. Objects
inside innerConeAngle are fully lit. Attenuation is then gradually
interpolated to 0 when approaching coneAngle using smoothstep.

innerConeAngle is set to be 70% of coneAngle if left undefined, so that
the coneAngle can be defined alone.

Task-number: QTBUG-77611
Change-Id: I9a1ee3c43746a06c583db00ac7aec403ae16dc3e
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>