Support referencing the same scene from multiple View3Ds
This means that we need to include the "layer" (which is basically a
per-View3D object now due to not having a true Qt 3D Studio style layer
concept anymore) in the key for the uniform buffer sets. A different
View3D rendering the same model in the same scene will often have a
different camera, meaning the uniform data is actually different from
other View3Ds.

This fixes the smaller problem (incorrect camera on the teapots) with
the view3d example.

Task-number: QTBUG-81594
Change-Id: Ib8700353f0d4bfbcd1d1fbd0544851449a5fc1e6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>