Enable multiple View3D items with rhi
Only do beginFrame - endFrame once per frame.

We can get away with pretending each View3D renders its own "frame" when
they all have a renderMode of Offscreen, but this breaks down with Overlay/Underlay
View3Ds with RHI where the prepare and render step are separated, and may be
interleaved (meaning call sequences like prepare, prepare, render, render are
perfectly possible).

Also remove some unused stuff.

With this patch, the view3d example renders close to as expected, except
that the PrincipledMaterial has to be changed to DefaultMaterial since the
prinicipled is not yet supported. The other unrelated thing is the
background of the Underlay View3D, where clearing is problematic. These
are issues to be tackled separately.

Task-number: QTBUG-81594
Change-Id: I05e2b767a8d854b91ba8384f6e56d01a571eff45
Reviewed-by: Andy Nichols <andy.nichols@qt.io>