Enable shadow mapping with RHI
Functional with all QRhi backends. (for Metal a qtbase fix is needed as well)

The depth bias (to reduce self-shadowing and such) is currently disabled since a
qtbase patch needs to integrate first. Once that is in place, and
re-enabled, it will be fully identical to the direct GL path.

The generated shader and rendering code should match the direct OpenGL path. There
is quite some amount of extra logic in place to accommodate for the potentially different
Y axis in images, and the different clip-space depth range (0..1 vs.
OpenGL's -1..1). Cubemaps also need extra care in this regard.

The MRT-based blur with the omnidirectional technique (the one with cubemaps) is
only done on systems where the max number of color attachments is at least 6 (so it
will be gracefully disabled on some embedded devices, such as the i.MX8 that only
provides thte GL/VK spec mandated minimum of 4).

Also adds a quad renderer which can (perhaps partially, perhaps fully) reused in
other techniques (like SSAO) later on.

Also also standardizes max light count to 15, max shadow count to 8, on all sides
(C++, shaders).

The default material key now includes the list of vertex attributes that
are present in the mesh the material is supposed to be used together with.

In the string form of the key this shows up as, for example:

vertexAttributes={position=true;normal=true;texcoord0=true;;;;}

In the example this indicates that attr_pos, attr_normal and attr_uv0 are ok
but other vertex inputs should not be declared in the shader code.

This means that using two model nodes with .mesh files where one has certain
attributes whereas the other does not, will now lead to generating different
shader code for the default material even if any other material setting is
completely the same. This way we can avoid declaring vertex inputs that are then
not connected to anything in the pipeline. The missing inputs are replaced with
shader code that assumes 0.0 for the value (which is what OpenGL silently does).

Also removes all sorts of unused code that are somewhat related to the purposes
of this patch, in order to avoid bothering with code that is not even use at all.

Task-number: QTBUG-81555
Change-Id: I4ecea4556a23b4c9b0247f4070d0db203bc00478
Reviewed-by: Andy Nichols <andy.nichols@qt.io>