Fix default position for QSSGRenderCamera
The default z-position for the camera was at negative 600, which happen
to work fine as we used to default to using left handed coordinates,
but since we're now using right handed coordinates both for the front
and back-end nodes, the default camera should be placed at positive 600.
This was discovered when using the Effect system, as it will set-up
a camera for off-screen rendering, which uses the default value.

Change-Id: Ie4e90f3ec2dec66f8b88db880f90f4c8cf99f0c5
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>