Fix scene manager node references
The root node, which is the parent of the scene manager, might get
destroyed before the rest of the nodes, which are still referencing the
scene manager, which causes reset when they try to dereference the scene
manager. Remove unnecessary referencing calls and add check for the scene
manager, which deletes the scene manager only if it is not referenced by
any nodes anymore.

Task-number: QTBUG-81840
Change-Id: Ic83c01aa123a2558e85d85db29e2b180c4bff06a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>