Add support for Z prepass
Can be enabled by setting depthPrePassEnabled to true.

We take a shortcut for now: displacement map usage disables the Z prepass,
as if depthPrePassEnabled was false. (anything that alters the geometry
implies having a dedicated variant of the Z prepass vertex shader, performing
the same adjustments as the real material would, and that's too much to ask for
now)

Custom materials are not yet supported, as usual, although most of the
prepare and render code will be shared regardless of the material type.

Also using this opportunity to fix up the TODOs and leftovers around the
rhiPrepare() and rhiRender() that are there since the initial prototype.

Task-number: QTBUG-81933
Change-Id: I7b6015531067244d694bd484792012402d9e0cfe
Reviewed-by: Andy Nichols <andy.nichols@qt.io>