examples: Make Underlay backgrounds identical
With QSG_RHI=1 the scenegraph always clears when it starts the renderpass.
It is no longer possible to circumvent this and
QQuickWindow::setClearBeforeRendering() has no effect with the RHI
rendering path, as documented. As a consequence,
SceneEnvironment.clearColor is not used when the renderMode is not
Offscreen.

Therefore, the simplest solution to get the same results in Underlay
View3Ds regardless of QSG_RHI is to use the same value both for
Window.color and for SceneEnvironment.clearColor.

Change-Id: Ie8040970ea32bea653babb47c1696148e425ed4a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>