rhi: shadows: Avoid recreating FBOs(or equivalents) on every frame
Running the lights example with the OpenGL backend of QRhi showed an
unreasonable high CPU usage on some systems. This is traced back to the
fact that we unconditionally call build() from addShadowMapEntry() for
some of the QRhiTextureRenderTargets. This is not ideal because
addShadowMapEntry() is called on every frame. Setting up a new native
native resource is unnecessary in the first place if the rt is non-null.
Move the affected code blocks inside the if (!rt) blocks. This will cause
the CPU usage jump back to reasonable values with GL as well.

Change-Id: I95b1c3f7eea7ab005442988e33c0a2e71f0168a0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>