Add support for screen space ambient occlusion with QRhi
Implies introducing support for generating a depth texture in a separate
render pass. This will be reused for other purposes later on.

The ao shader is modified to have manually unrolled loops to help HLSL.

With OpenGL, SSAO is only available on OpenGL ES 3.0 or OpenGL 3.2+. (no
texelFetch in earlier GLSL versions) This we cannot currently detect
in Quick3D without adding the corresponding feature flag to QRhi. So the
check for this is disabled until the qtbase patch merges.

Comes with the usual set of cleanups and removal of unused code.

The one Lancelot test case gives identical (in fact better, since 5.15
is a bit broken) results compared to the direct OpenGL path. SSAO is to
be exercised a bit more once a suitable scene (and models) are created.
That is left as an exercise for a separate patch.

Change-Id: Iaddbc90d7549a1369560bcc157adda5ac170b853
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>