Add Lancelot test for material-specific IBL
The RHI port had a bug that affected materials with a lightprobe
when there was no global lightprobe in the scene (fixed in commit
cf76854294e1d8d97f1d0010133d7d7f5b378eb1). The existing test scene
Imagebasedlighting.qml did not catch this, since it has a lightprobe
in the scene environment.

Change-Id: I329273335b50718ac42acbbccf94a7d447d3d73a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>