Fix the render context look-up map
Don't store the ref directly, we already own the ref to the render
context in the scene manager, and when the scene manager goes out of
scope we the list is irrelevant and we don't need, or should, manually
release the map. Also, there's no need to store the render context
in the scene manager, it's already part of the context interface.

Change-Id: Ida5cf4f23fb06ae12d948252688c85b4741584cf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>