Add support for supersample antialiasing with RHI
Supersample antialiasing is based on rendering the Offscreen View3D into a
texture larger than the desired final size. The multiplier is typically
1.5, 2.0, or similar. This also implies making the depth stencil buffer
match the increased size.

Once the renderpass is done, an additional pass is neede to get the content
from the larger texture into the View3D's smaller backing texture. With
OpenGL (and Vulkan) this could be achieved by glBlitFramebuffer
(vkCmdBlitImage) but there is no equivalent operation in Metal or D3D.
(note that we rely on scaling and linear filtering, both of which are
extras that are not supported by the ordinary texture copy operations)

Therefore, we just draw a quad, sampling the throwaway ssaa texture, and
rendering into the View3D's backing texture.

Change-Id: I1960247cc9a5d6de421a7471c8ae8dfc705eb542
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>