Fix LookAt API errors
The math was a bit busted, and QML needed to QQuick3DNode to be
registered.

Quaternion.lookAt(...)'s function signature changed a bit to accommodate
some default arguments.  The return of this function is the relative
transformation that needs to be made to get the correct lookat value, so
to use it you either need to use the default forward/up vectors to get
the global rotation, or the current forward/up vectors to get the
relative change.

Implement QQuick3DCamera::lookAt() using Euler angles to avoid
rolling the camera.

Fixes: QTBUG-83620
Change-Id: I48640f8b09e13b2b3f669c6d65aa7c1f9af9730f
Reviewed-by: Christian Strømme <christian.stromme@qt.io>