Make sure context, rhicontext, etc. get torn down when exiting
There now asserts in place that will trigger when one holds on to
the QSSGRef for the QSSGRhiContext or the QSSGRenderContextInterface.

There are some rhi resources leaked in the more complex apps, but that's
not caused by not destroying the context and others.

Simple apps like hellocube can now exit cleanly (when it comes to graphics
resources at least) with rhi as well.

Change-Id: Id35c17abadff53b3468b0164b1eb41cdc065191d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>