Do not leak QRhiTextures
Introduce the most basic bookkeeping: register the textures to
QSSGRhiContext and it takes ownership and that's that.

This will probably become smarter later (as we will need to introduce
something that allows destroying textures that are no longer needed
after dynamic scene changes - same goes for buffers as well). But this
will do for now, it allows reaching 0 leaked rhi resources in the examples.

Change-Id: Ia949949072ef3acd5324ae496662bbdc5a0a796d
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>