rhi: Enable custom materials with lighting and shadows
...as much as possible.

Introduces custom material support for lighting, light probes, and shadow
mapping. Also blending and culling modes.

Also handles uEnvironmentMappingEnabled correctly for custom materials
now. This means that setting it to false actually disables sampling the
environment map texture.

It is worth noting that the built-in custom material shaders are not
prepared and, as of now, should not be expected to handle lighting: it
seems they were not created with lighting and shadow maps in mind. This
means they will not correctly participate in lighting and shadowing, even
when environment mapping is disabled. It is out of scope for this patch to
handle this further.

Task-number: QTBUG-82245
Change-Id: Id70714ef45c8e4b9824aaa626ca7b5ad410a7887
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>