Generate only one set of shaders for OpenGL too
Letting it generate 3 GLSL versions of each shader is both a waste of time
and is also a problem with custom materials that tend do use some modern
features like the uint type. For such a shader attempting to generate a
GLSL 100 es version will fail. To prevent this, try to make an educated
guess based on the default surface format (typically set by
idealFormat()).

Change-Id: Id4b4540e9b2991821abb7844bdf36da54299cce8
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>