Remove 64-bit QImage conversion warning
This is not planned to be supported in the foreseeable future, so the conversion
will always be necessary if the input is, for example, a 16 bits per component PNG
file.

If we gain support for RGBA16 texture formats in the graphics stack in the
future (a low priority feature for now), the conversion can be removed.
But until then, there is no point in showering warnings all the time.

Change-Id: Ie6f9b977a6f9ec5d6ae18123eeb7053ce2ad4491
Reviewed-by: Christian Strømme <christian.stromme@qt.io>