Fix type API
The type() function cannot be a virtual function, we use this function
several places for run-time type information, and most critically in
the destruction of the objects, so it needs to be stable!
In the base class the function was pure virtual, which meant we would
in cases call the pure virtual function and suffer the consequences.

Change-Id: I530744cf9ff818309b4a8b5141ff6567c25e6914
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>