Don't expose the scene manager in the QQuick3DObject class
There's no use-case for this as far as I can tell, and handing out a
raw point from a QSharedPointer is far from ideal.
The scene manager can be retrieved through the private implementation,
if needed. Also removed the sceneManagerChanged() signal, it's not
used and we already have infrastructure to notify scene manager changes.

Change-Id: Ie0c4f72f60631ff2dd697be22bb3d9251af45f1f
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>