Fix material handling in QQuick3DModel
Make sure the material has a scene manager set. In most cases
it's sufficient to set the parent item of the material, if one is not
set already. If the material has no parent item set, check if it has
an immediate hierarchical parent that can be used, this is normally
the case for inline materials, materials declared at some random
place in the scene or properties. There are other cases where the
immediate parent is of a non-suitable type and we should handle
that as well, by manually updating the material with the model's
scene manager through the ref-/derefSceneManager() functions.

Fixes: QTBUG-82276
Change-Id: Id0d758235376520374c288712dd220e901a76aeb
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>