Add support for conditional sampler declarations
So something like

  { "type": "sampler2D", "name": "shadowMaps[4]", "condition": "QSSG_ENABLE_SSM" }

should expand to

  #if QSSG_ENABLE_SSM
  layout(binding = 12) uniform sampler2D shadowMaps[4];
  #endif

whereas until now the #if was missing because it was only added for
uniform block members, not for samplers.

There is also a change that affects uniform block members as well: the
correct thing to do is to use #if not #ifdef. Feature set flags are
always written out, such as #define QSSG_ENABLE_SSM 0 so testing this
with #ifdef is wrong (because it is always true).

Task-number: QTBUG-82245
Change-Id: I9679be28640f6a0edfdf32d041140553503ec8d8
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>