Enable, in theory, SSAO for custom materials
In practice the ambient occlusion factor calculated is not used in any
of the built-in custom materials, so SSAO has no effect (even though
enabling it will generate the AO map). This is out of this patch's scope
to solve. What this ensures is that the textures are correctly exposed.

Also note that SSDO is not supported in the code base (it never was, and
has no public API to enable it), so we will not bother with the cbAoShadow
uniform block that would only be used by SSDO.

Change-Id: Ic4cb90f52276ecf364c20fffd1c9355702773c5d
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>