Fix crash on exit
Calling cleanupNodes() in the destructor of the scene manager will
cause a crash, as some of the nodes will try to release their
GL resources when theres no valid context to be used for that.
This change is a simple workaround, a more appropriate
change would be to get the destruction correct and clean-up properly,
but that's going to be a bit more involved, so for now it's better
to accept bad clean-up at exit then a crash.

Fixes: QTBUG-82427
Change-Id: I7a2ad8850c27f5544547f6bf207755dd7a1b9213
Reviewed-by: Andy Nichols <andy.nichols@qt.io>