CMake: Handle system assimp correctly
Make sure to run the config test and only use system assimp
if the compile tests succeeds.

Make sure to mark the Find module as not found when the system
assimp library is not found.

Work around the broken assimp Config file which only sets
IMPORTED_LOCATION_RELEASE on non-Windows platforms, without setting
IMPORTED_CONFIGURATIONS, which means that if Qt is configured in Debug
mode, the build will fail due to not finding an appropriate library
file.

Adapt the config.test to find and link against system assimp.

Change-Id: I05299dce2a5589a95c9300c502fc303ba475a958
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>