Add rhi support for QQuickItem children in QQuick3DObject
This comes with the caveat that the depth prepass, when enabled, will be
implicitly disabled whenever a 2D item child is added to a 3D node.
Supporting the Z prepass would involve adding support for drawing these
quads into the depth buffer similarly to how we draw the model geometry,
but this is left as a future exercise for when we decide if the cost of
implementing and maintenance is worth it for an opt-in, likely-never-used
rendering feature.

The leaking of objects (likely QSGLayer objects, which in turn lead to
not releasing graphics resources on exit) from Item2D is also not solved
here as that is likely present in the 5.15 code base as well. To be
handled separately.

Other than that, the quickitem example and the related Lancelot tests
should give identical results.

Task-number: QTBUG-83469
Change-Id: I33a63bd1b47ae34ce66a120b267386b8f2233bf9
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>