Set default Shadow filter factor to 5 rather than 35
Applying a value of 35 created a very diffuse shadow with many artifacts
rather than simply bluring the edges a bit to create a softer edge to
shadows.  A value of 5 will provide significantly harder shadows, but
still with some softness (not pixel aliased, but will show polygon edges
for low-poly meshes). This is a more sensible default that will look
better in most scenes.

This is a behavior change, but since shadows generally didn't look
correct before, most "real" scenes manually adjust the shadowFactor
already.

Change-Id: Ida38d6c581ee66005ee262b31171fbf2446e880e
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>