Adjust RenderModeUnderlay Lancelot test
Similarly to what we did with the examples: with the RHI-based rendering
path the View3D's clearColor has no effect when the renderMode is not
Offscreen. So just set that color to match the QQuickWindow's color.
That way the end result is the same regardless of who does the clear.

Task-number: QTBUG-83801
Change-Id: I3f63da7e711b849ad015d7bd824c7a40dedf884f
Reviewed-by: Andy Nichols <andy.nichols@qt.io>