Fix using effects and MSAA together
Use the new argument in QRhi::newRenderBuffer().

This hint is used by QRhi backends to allow creating a texture
with the same format as the one the multisample resolve is targeting.

Without effects everything works because then we have an RGBA8->RGBA8
setup, but once effects are involved, the main render target contains a
multisample resolve from RGBA8 to RGBA32F. This is not supported
(outside OpenGL, and maybe Metal, and it may sometimes work with Vulkan,
although it is a validation error), and is also wrong with regard to
precision and quality. Avoid all this by making sure we render to a
(multisample) RGBA32F image in the first place.

Task-number: QTBUG-83707
Change-Id: Icdb14d45ede134f56706c4b42ca3531693098941
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>