Skip non-opaque objects when generating the depth texture
The 5.15 code base looks quite confusing in this regard, there must be
some history behind this there because it looks the semi-transparent objects
were meant to be included originally, but this then got disabled in a somewhat
confusing manner.

As it stands, it skips the semi-transparent object list in the depth texture.
To remain compatible, follow suit on the RHI-based rendering path. It is
simple because it only involves not passing the semi-transparent list to a
function.

Task-number: QTBUG-83874
Change-Id: Iddb417b0abad65b2e18f15499d0b7741b82ce631
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>