Avoid stale QRhiRenderPassDescriptor objects in the pipeline cache
When changing AA settings, we do drop a number of objects, including
render targets and related resources. This is a low frequency event, so
removing associated m_pipelines entries should be fine. (if someone
frequently changes AA settings, they will have much bigger performance
issues than the ones arising from not reusing a pipeline object)

We follow the typical deferred-release pattern that is employed in
various places by QRhi, and in some ways Quick3D as well:

When an object that is part of the key to a pipeline cache entry becomes
invalid, we remove the entry right away, but only destroy the object
(the QRhiGraphicsPipeline in this particular case) upon the next
beforeFrameBegin() signal from QQuickWindow.

So even if the invalidation function is called in the middle of preparing a
frame, with the QRhiGraphicsPipeline referenced on the (under the hood not yet
recorded) QRhiCommandBuffer, nothing breaks since the pointer stays perfectly
valid for entirety of the preparation of the that frame.

This matches how QRhi backends, like the Vulkan one, manage native
graphics objects: it is QRhi::beginFrame() that issues the vkDestroy*
of calls for objects that are queued for releasing and are deemed safe to
be released.

The pattern can later be extended other objects, like for samplers,
textures, uniform buffers, and SRBs, once a more dynamic method of
managing the lifetime of mesh and material related sources is introduced.

Task-number: QTBUG-83883
Change-Id: I959c5e541122a436d1220d862280b740ab9a714f
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>