Use new per-target compile mode for prog.aa and ssaa shaders
These had separate _flip variants because dynamic branching in the
shader based on uniforms was deemed unnecessary for these small shaders.

There is now a better way: by passing -p to qsb one can now ifdef based
on QSHADER_language macros. So we can have one .frag file and one
.frag.qsb because the GLSL/HLSL/MSL/SPIRV in the .qsb is magically
compiled with a different set of QSHADER_ macros defined. So it all just
works.

This comes with the caveat of making shading language -> graphics API
assumptions. But here the benefits outweigh the risk of theoretical
future problems. (for complex shaders, such as the materials, this
approach is not recommended because there there are already many different
cases of passing in flags from C++ in uniforms, but for the simple, static,
built-in shaders it is very useful)

Change-Id: I199c631ea21e10c24e4daf1ecf05e1fcb5de9512
Reviewed-by: Andy Nichols <andy.nichols@qt.io>