Render 2D items without and FBO when possible
This changes how 2D Qt Quick items are rendered inside a 3D scene.
Instead of creating an offscreen surface, a new Qt Quick 2D renderer
is created to renderer the 2D content in the same pass as the
3D content.

There are some behavior changes when 2D Items are used in this manner.
Instead of generating a Quad mesh that is centred on the Node's origin,
2D Items will now be placed with their top-left corner in the node's
origin. To recreate the previous behavior use: anchors.centerIn: parent

There is also a hidden "contentItem" now that is created when 2D items
are added as children of a Node. All of this should make for a more
smooth experience embedding 2D content in 3D.

Change-Id: I9060a291c2652818da7742249f9e0255a1b34356
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>