Reduce QQuickWindowPrivate dependency in scenerenderer
This, like the original code, suffers from not being compatible when the
QQuickWindow has its rendering redirected in combination with
QQuickRenderControl (there is no swapchain in that case).

However, the solution to that lies along what's shown here, the
QSGRendererInterface will eventually expose more resources. For now
as a first step let's just decouple from QQuickWindowPrivate.

Change-Id: Ib3239d7a96217201ac0ce08fa24c8bc03c725a9b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>