Move generated uvcoords state to QSSGRenderableImage
This state was stored in the shader generator, but storing it in the
QSSGRenderableImage class gives us more flexibility, as there's no need
to keep a separate table to track each image or adjust it to match the
different types we at any time support. The image type enum was also
moved into QSSGRenderableImage's scope, for clarity, and the size was
change to not affect the size of QSSGRenderableImage. Note that the
state of the uvcoords generation is naturally reset on each frame.

Change-Id: Ie4d34dc7901063c48ebf0686bcf77ba83c866778
Reviewed-by: Andy Nichols <andy.nichols@qt.io>