Refactor the QSSGMaterialShaderGenerator class
We want this code to be self contained as much as possible, as we want
to decouple it from the to the point where it's usable out side the
engine itself. With this change there's no longer a relationship
between the custom material generator and the "default" generator.
The generator itself also contains no state, which should make it
easier to interface with. While most of the state the generator
kept was complete nonsense, there was some caching mechanisms to speed
up uniform name look-ups  etc., those problems have been taken into
consideration (should hopefully be even faster). The generated
sampler names are also now more sensibly named now (based on the type).

While the generator now is less coupled with the rest of the engine,
further work is needed to decouple the pipeline and its sub-components
to make the whole infrastructure usable outside the engine.

Task-number: QTBUG-84770
Change-Id: I3b364c02ff0729ab45a835df8b39601f65f24a35
Reviewed-by: Andy Nichols <andy.nichols@qt.io>